requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.7.1
python-multipart==0.0.9
librosa==0.10.2
numpy==1.26.4
soundfile==0.12.1
pydub==0.25.1
pyloudnorm==0.1.1
openai-whisper==20231117
deep-translator==1.11.4
aiofiles==23.2.1
